Input data from CPP output into a list of data frames

Input data from CPP output into a list of data frames

parseSimulationOutput(filename)

Arguments

filename
The filename of the output generated by the simulation

Value

A list containing

params

the parameters used to run the simulation

sequences

a data frame of the sequences over time and their similarity to the initial sequence

step - timestep in the simulation

realTime - time since the start of the simulation (in millions of years)

sequenceId - the unique ID of the sequence (to track it over time); initial sequence has sequenceId 0 to (numInitialCopies-1)

parentMain - the unique ID of the sequence this burst from; (-1 if nothing)

parentOther - the unique ID of the sequence its parent recombined with; (-1 if nothing)

distanceToInitial - the distance to the initial sequence

isActive - whether or not the sequence is capable of transposition

pairwise

a data frame of the pairwise similarity of sequences over time

step - timestep in the simulation

realTime - time since the start of the simulation (in millions of years)

sequenceId1 - an ID of a sequence present at the time

sequenceId2 - an ID of a different sequence present at the time; not all pairs are given - that is, for sequences a and b, either (a, b) or (b, a) is present as a row but not both

distancePairwise - the distance between the two sequences

familyRepresentatives

a data frame of the family representatives for the retrotransposons over time

step - timestep in the simulation

realTime - time since the start of the simulation (in millions of years)

familyId - the unique ID of the family representative (to track it over time)

creationTime - time this family representative was created (in millions of years)

sequenceId - a sequence belonging to that family (sequences belonging to a family are listed as separate rows, and all sequences belonging to that family are listed)

familyPairwise

a data frame of the pairwise similarity of the family representatives for the retrotransposons over time

step - timestep in the simulation

realTime - time since the start of the simulation (in millions of years)

familyId1 - an ID of a family representative present at the time

familyId2 - an ID of a different family representative present at the time; not all pairs are given - that is, for sequences a and b, either (a, b) or (b, a) is present as a row but not both

distancePairwise - the distance between the two family representatives

Examples

if (FALSE) {
data <- parseSimulationOutput('simulationOutput.out')
}